[feat] integer 的id全部变更为 Long

+++ ruoyi-admin/src/main/java/com/renxin/consultant/controller/wechat/WechatConsultantPayV3ApiController.java
@@ -223,7 +223,7 @@ public class WechatConsultantPayV3ApiController extends BaseController {
     @PostMapping("/wechatRefund")
     public Map<String, Object> wechatRefund(String transaction_id) {
         //@TODO demo中先写死的一些参数
-        Integer userId = 1; //先写死一个用户id
+        Long userId = 1L; //先写死一个用户id
         BigDecimal amount = new BigDecimal("0.01"); //先写死一个退款金额 单位：元
         String reason = "支付demo-退还订金"; //先写死一个退款原因 这是可选的
