+++ ruoyi-admin/src/main/java/com/renxin/web/controller/gauge/PsyGaugeBannerConfigController.java
@@ -64,7 +64,7 @@ public class PsyGaugeBannerConfigController extends BaseController
      */
     @PreAuthorize("@ss.hasPermi('banner:config:query')")
     @GetMapping(value = "/{id}")
-    public AjaxResult getInfo(@PathVariable("id") Integer id)
+    public AjaxResult getInfo(@PathVariable("id") Long id)
     {
         return AjaxResult.success(psyGaugeBannerConfigService.selectPsyGaugeBannerConfigById(id));
     }
@@ -108,7 +108,7 @@ public class PsyGaugeBannerConfigController extends BaseController
     @PreAuthorize("@ss.hasPermi('banner:config:remove')")
     @Log(title = "测评banner配置", businessType = BusinessType.DELETE)
 	@DeleteMapping("/{ids}")
-    public AjaxResult remove(@PathVariable Integer[] ids)
+    public AjaxResult remove(@PathVariable Long[] ids)
     {
         try {
             int res = psyGaugeBannerConfigService.deletePsyGaugeBannerConfigByIds(ids);
